// src/generated/models/Pm_financialreportconfigsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pm_financialreportconfigspm_charttype = {
  1: 'BarChart',
  2: 'LineChart',
  3: 'AreaChart',
  4: 'StackedBarChart'
} as const;
export type Pm_financialreportconfigspm_charttype = keyof typeof Pm_financialreportconfigspm_charttype;
export const Pm_financialreportconfigspm_groupby = {
  1: 'Project',
  2: 'CostCatagory',
  3: 'FiscalPeriod',
  4: 'FundingSource'
} as const;
export type Pm_financialreportconfigspm_groupby = keyof typeof Pm_financialreportconfigspm_groupby;
export const Pm_financialreportconfigspm_hierarchylevel = {
  1: 'Global',
  2: 'Portfolio',
  3: 'Programme',
  4: 'Project'
} as const;
export type Pm_financialreportconfigspm_hierarchylevel = keyof typeof Pm_financialreportconfigspm_hierarchylevel;
export const Pm_financialreportconfigsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pm_financialreportconfigsstatecode = keyof typeof Pm_financialreportconfigsstatecode;
export const Pm_financialreportconfigsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pm_financialreportconfigsstatuscode = keyof typeof Pm_financialreportconfigsstatuscode;

export interface Pm_financialreportconfigsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  pm_categoriesfilter?: string;
  pm_charttype?: Pm_financialreportconfigspm_charttype;
  pm_description?: string;
  pm_financialreportconfigid: string;
  pm_fiscalyearsfilter?: string;
  pm_groupby?: Pm_financialreportconfigspm_groupby;
  pm_hierarchylevel?: Pm_financialreportconfigspm_hierarchylevel;
  pm_ispublic?: boolean;
  pm_name: string;
  pm_selectedcolumns?: string;
  pm_selectedfilters?: string;
  statecode: Pm_financialreportconfigsstatecode;
  statuscode?: Pm_financialreportconfigsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pm_financialreportconfigs extends Pm_financialreportconfigsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pm_charttypename?: string;
  pm_groupbyname?: string;
  pm_hierarchylevelname?: string;
  pm_ispublicname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}